requirements.txt
etcd3==0.12.0
grpcio==1.64.1
joblib==1.4.2
keras==3.3.3
protobuf==3.20.3
scikit_learn==1.5.0
tensorflow==2.16.1
pandas
grpcio
grpcio-tools==1.48.2